requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.3
Django==3.2.8
djangorestframework==3.12.4
install==1.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pytz==2021.3
regex==2021.10.21
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
